fold review: union-race posture + coverage test split (issue #200)
